+++ js/src/forum/states/QualityState.ts
@@ -1,0 +1,11 @@
+export default class QualityState {
+    name: string;
+    url: string;
+    type: string;
+
+    constructor() {
+        this.name = "";
+        this.url = "";
+        this.type = "normal";
+    }
+}