+++ src/Api/Serializer/BasicPostSerializerAttributes.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Nearata\EmbedVideo\Api\Serializer;
+
+use Flarum\Api\Serializer\BasicPostSerializer;
+use Flarum\Post\Post;
+
+class BasicPostSerializerAttributes
+{
+    public function __invoke(BasicPostSerializer $serializer, Post $post, array $attributes)
+    {
+    }
+}