+++ src/ForumSettings.php
@@ -29,7 +29,7 @@ class ForumSettings
             'embedVideoTheme' => (string) $this->settings->get('nearata-embed-video.admin.settings.options.theme'),
             'embedVideoLogo' => (string) $this->settings->get('nearata-embed-video.admin.settings.options.logo'),
             'embedVideoLang' => (string) $this->settings->get('nearata-embed-video.admin.settings.options.lang'),
-            'embedVideoModal' => (bool) $this->settings->get('nearata-embed-video.admin.settings.options.modal_enabled')
+            'embedVideoModal' => (bool) $this->settings->get('nearata-embed-video.admin.settings.options.modal_enabled'),
         ];
     }
 }