+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- bump php to 8
+- bump flarum to 1.7
+- update dependencies
+
 ## [3.2.1] - 2022-09-09
 
 - Fix BBCode not working with alphanumerical ids.